SEDP ignores dispose messages with security
Problem
-------

The code path for unsecure discovered readers and writers in SEDP
checks that the topic is indeed not secure using the topic name.
However, the topic name is not populated for unregister/dispose.
Consequently, unregister/dispose messages are dropped.

Solution
--------

Check for sample data, i.e., the topic name should be populated,
before the security check.