Sedp needlessly defers work
Problem
-------

Every input message received by SEDP is placed in a JobQueue.  This
queueing is unnecessary.  Association complete messages are also
queued.  This was to avoid a deadlock but deadlock can be avoided by
checking if the writer/reader is reliable.  Finally, removal of a
subscription/publication queued a job to remove it from the BIT.  This
queueing is unnecessary.

Solution
--------

Remove the queueing.