RTPS Transport does send durable data for first ACKNACK
Problem
-------

The "association complete" callback for a reliable RTPS writer is
called when the writer receives the first acknack for a reader.  If
the writer is durable, it causes the writer to send the durable data
for the new reader.  This is currently done and the end of the ACKNACK
processing.  Durable data is sent in response to an ACKNACK.
Consequently, a reader must send one ACKNACK to enqueue the durable
data and another to retrieve it.

Solution
--------

Invoke the callback earlier so the durable data will be available for
sending in response to the first ACKNACK.