RTPS reliable writes are not scalable
Problem
-------

For each message, a (reliable) RtpsWriter loops through all readers to
check that none are expecting durable data for the purpose of
inserting a gap if a sequence number is skipped.

Solution
--------

Gaps should only be sent when the maximum sequence number increases.
When this happens, all readers can receive a gap.