Some routers confused over STUN traffic
Problem
-------

When using the RtpsRelay and ICE, the SPDP port will send STUN packets
to both the SPDP and SEDP port of the relay.  In practice, this
confuses the router and the observed behavior is that all traffic from
the relay's SPDP port was dropped.  The SEDP and DATA channels when
using the same router are not affected.

Solution
--------

Add configuration so that SPDP will send STUN to the SPDP port of the RtpsRelay.