RtpsRelay stats don't distinguish directed and undirected gain
Problem
-------

When a client sends a message to the RtpsRelay it may get forwarded to
N other clients.  The number N is the gain and is an important
statistic for the RtpsRelay as it will be more performant for small N
and can be used to diagnose potential scalability problems.

Messages where all submessages are prefixed with INFO_DST must be sent
to all named destinations (if they exist).  Otherwise, the relay must
forward the message using the association table.  When applied to
gain, the former is gain for directed messages and the latter is gain
for undirected messages.

Currently, the RtpsRelay does not distinguish these two cases.

Solution
--------

Add statistics for the maximum directed and undirected gain observed
over a statistics interval.