RTPS reliability has a race condition
Problem
-------

A common pattern in testing and perhaps applications is for a reliable
writer to wait until it has matched a certain number of readers.  The
common expectation is that any sample written after this point will be
received by the reader.  This does not happen due to a race condition
between heartbeats and data.

The start of the "nackable" range is determined by the first received
heartbeat or first received data sample.  If a data sample is received
first and it is not the first sample, then the reader will skip over
all previous samples.

Association is complete on the writer side when it receives an acknack
from the reader.  In the scenario described, the writer could write
all of the data and call `wait_for_acknowledgements` before the reader
has received anything.  If the reader receives a heartbeat first, it
might end at the final sample produced by the writer and the reader
will end up not receiving any data.

A solution where the acknack is delayed until the reader receives a
heartbeat does not support cases of assymetric discovery, i.e., the
reader has to be able to send an acknack to initialize rediscovery.

Solution
--------

Delay association of the writer until it receives an acknack that
proves the reader has received a heartbeat.  Such a heartbeat would be
final, ack something, or have a non-empty bitmap.