Association releated features broken for local RTPS participants
Problem
=======

1. Waiting for publication matched status fails for local RTPS
participants.  This was caused by the the discovery and transport code
assuming that local RTPS reader/writer pairs are automatically
associated.  This would allow a writer to violate the new reliablity
protocol and send samples before the reader was prepared to receive
them.

2. Suppose a reliable, durable writer writes samples before any reliable
reader is associated.  The heartbeat for this writer is still [1,0)
since no data has actually made it to the datalink.  Suppose two
readers are discovered at the same time, one durable and one not.
They will both receive a [1,0) heartbeat which is not necessarily
wrong.  After receiving this heartbeat, the readers will acknack
causing association complete and replay of durable data.  The durable
replay will advance the maximum sequence number to the correct value
but since it is directed, only the durable reader will be viewed as
lagging.  The non-durable reader is still leading at zero.  Writing
new samples will advance the sequence number but the leading/lagging
logic will never make the non-durable reader a lagger because the
sequence number it is leading at is still zero.

Solution
========

1. Remove `association_complete` as a concept in discovery.  Discovery is
allowed to say was associations should or should not exist but it
should not participant in the determination that those associations
are complete meaning that data written by a writer will be send to a
reader.  This means that it is the responsibility of the transport to
invoke `transport_assoc_done` when association is complete.

2. Initialize RtpsWriters with a starting sequence number.