Simplify synchronization for TCP transport
The flow when establishing a connection with the TCP transport is now
as follows:

When a connection is made to or found on the reader side, the reader's
callbacks are executed to make the framework ready to receive messages
and the ack request is sent to the writer to indicate that the reader
is ready to receive messages.

When the writer receives the ack request, it's callbacks are invoked
and it can start writing data.