Reliability in multicast transport is broken
Problem
-------

The association of two reliable multicast transports can finish before
the reader is associated with the writer through discovery.  This
means a writer may start writing data with the expectation that the
reader will receive it.  However, the association may not exist on the
reader side so it will not receive the data.

Solution
--------

Extend the SYN+ACK protocol to include reader-writer pairs.  The
passive side will only ACK a pair when the association has been made
known through discovery.  The SYNACK packet can be interpretted as
association complete and the callback on the writer is invoked.