Thrasher test lacks durability scenarios
Problem
-------

The Thrasher tests create T reliable writers in T participants and 1
reliable subscriber.  The writers write N messages and the reader
checks that all messages are received.  The test stresses
discovery (especially RTPS) and reliability including features like
waiting for acknowledgements and waiting for match status.  Durability
is implicitly tested through discovery.  The test could be used to
diagnose problems with durability if it were extended.

Solution
--------

Add flags to enable durability.  When the writer is durable, it either
1) waits for a match and then writes data or 2) writes data and then
waits for a match.