Replaced call to cancel_and_wait in RevokePermissionsTask dtor.
The removal of the call to cancel_and_wait was suggested as a fix to the failing OpenDDS_AccessControlTest test.  Originally, that change was part of a larger PR (2024).  This change did not fix the test, which still fails with the same error: "SEH exception with code 0xc0000005 thrown in the test fixture's destructor".