Delivering held data in RTPS is asynchronous
Problem
-------

The decision to deliver held data results in an event handler being
scheduled with the reactor.  This puts the system in a state where it
may deliver samples received after the event handler is scheduled but
before it executes.  This results in out-of-order delivery for
reliable readers.

Solution
--------

Use an RAII job to deliver held data immediately.