Durable writers send non-directed gaps
Problem
-------

A durable writer should never send a non-directed gap unless it is a
gap for all readers because a durable reader could lose data.  This
currently happens in two places in the RTPS transport corresponding to
the two places that gaps are generated.

The first is inserting a gap due to a discontinuity in sequence
numbers from the TransportClient (DataWriter).  The condition of not
sending a gap if skipping sequence number 1 is not sufficient.  This
condition was inserted to handle start up situations where data is
being written while the first reader is associating.  This situation
of the first reader associating may happen multiple times.

The second is gaps generated in response to an acknack.  This logic is
generally unsound as the decision to send undirected gaps is based on
whether a subset of the current readers is expecting durable data.
This problem is not addressed in this commit.

Solution
--------

Don't insert undirected gaps for durable writers when the sequence
number jumps.