Pending RTPS writers/readers orphaned
Problem
-------

When a reliable TransportClient using RTPS associates with a reliable
peer that peer is put in a pending state.  If the client disassociates
with the peer while it is still pending, then it is not cleaned up in
the datalink.  Consequently, the remote peers linger and consume
resources trying to associate.

Solution
--------

Inform the datalink that the pending peer should be disassociated.