Application participant sends non-directed heartbeats
Problem
-------

The reliable SPDP writer in the application participant of the
RtpsRelay can send undirected heartbeats that will be dropped by the
relay.

Solution
--------

Force directed heartbeats for the reliable SPDP writer.  This is a
work-around that may be removed in the future if the reliable SPDP
writer is disabled in the application participant.  Otherwise,
additional configuration plumbing will be needed to accomplish the
same result.