Durable RTPS writers may send gaps
Problem
-------

The `send_and_gather_nack_replies` function attempts to not
send (undirected) gaps if a durable reader is detected.  This is
flawed in two ways.  First, a durable writer should never send an
undirected gap because it cannot know when a potential reader will
come online.  Second, the existing logic only checks readers that need
a reply instead of all of the readers when making the determination.

Solution
--------

Re-write `send_and_gather_nack_replies` so that naks for directed data
receive a directed reply (gap or data) and naks for undirected data
receive receive an undirected reply.  Directed data includes durable
resends and messages from the participant volatile secure writer.  The
code for responding to naks for durable data will/should not be
executed until the logic that processes acknacks is updated.