Lost SPDP prevents secure discovery
Problem
-------

RtpsRelay deployments send directed SPDP announcements with a very
long period.  Furthermore, they send one (and only one) announcement
when a client is first discovered.  If that announcement is lost,
discovery will certainly stall and fail.

Solution
--------

Send an SPDP announcement with an authentication resend.  The client
will send SPDP to the relay frequently so the relay should discover
the client in a timely manner.  The relay will start authentication
resends, which happen with a very small period, so the client will
eventually discover the relay.

The non-secure case is not handled in this commit.