Reliable RTPS readers don't know where to begin
Problem
-------

For reliability to work (see below), the reader must have some
commitment from the writer that it will receive all samples past a
certain sequence number.  That sequence number defines the beginnning
of the reliable range.  For durable reliable readers, the reliable
range begins at one.  Non-durable readers have a choice.  The current
implementation picks the lastSN + 1 offered by the writer based on the
first heartbeat it receives.

However, given network delay, reordering, etc., the minimum sequence
number the writer has for reliable resend may be greater than the
sequence number picked by the reader.  In this case, the reader will
receive a gap to catch up.  If the reader has received some data after
the beginning of it's reliable range, then that data will be delivered
(still in order but stopping, otherwise, no gap).  If there is gap,
the application reader uses  monotonically increasing counter, then it
detect a missing sample.

Solution
--------

The writer and the reader have to agree on the beginning of the
reliable range.  The writer has to offer a starting sequence
number (through a heartbeat) and the reader has to accept it (through
an acknack).  In order to offer, the writer must not allow changes to
the send buffer until the reader has accepted the offer.  This is
accomplished by not removing data from the send buffer while offers
are outstanding, i.e., there are pre-associated readers.  The reader
has to accept the sequence number (or a subsequent sequence number)
offered by the writer.  To do this, logic was added to initialize the
reader on a directed heartbeat.  This breaks interoperability.

Note
----

There are a couple of use cases where agreement on the beginning of
the reliable range is important.  First, tests and applications often
use the publication matched status as a signal that every subsequent
sample will fall in the reliable range.  Thus, there is a guarantee
that subsequent samples will be delivered to the reader.  Second, the
participant volatile secure (PVS) writer for DDS Security is reliable
but not durable.  Thus, the PVS writer must defer writting samples
until the reliable range has started.  Thus, the PVS writer uses the
same technique as publication matched status.