Volatile reliable RTPS writers/readers don't start at the end
Problem
-------

The starting sequence number offered by the writer is not stable.

Solution
--------

Associate immediately on the writer side so that samples will
accumulate in the send buffer.  Record the state of the
writer (max_sn) at the time of association so that it can be offered
to the reader via the preassociation heartbeat.  The reader will be
offered an empty range starting at one past the heartbeat first.