Improve Secure Participant User Data
Building on #2066, if `SecureParticipantUserData` is active only allow
publishing to the Participant BIT if the data is valid. Since this
behavior change is on the receiving side, it means the
`SecureParticipantUserData` setting is required in all participants for
this to work properly, not just on the sending side as before.

Also added two cases for this to the security attributes test.