Participant BIT and location BIT not populated
Problem
-------

A secure remote participant was never published to the participant BIT
and, consequently, the location topic was never published.

Solution
--------

When security is enabled, publish the participant in the BIT when the
secure SPDP message arrives.  Add the ability to publish the location
topic even when there are no updates.