Durable replay not complete for nack
Problem
-------

Before, the datalink thread was responsible for invoking the callback
that replayed durable data.  Thus, durable replay was done when the
call returned.  Now, the discovery and/or user thread is responsible
for replaying durable data and it might not be finished by the time
the first nack arrives.  This results in a request that is gapped
leading to missed durable data.

Solution
--------

Prevent further nack processing by clearing the requests if the
durable data is not complete.