Reauthentication fails
Problem
-------

The main challenge in reauthentication is re-keying the volatile
reader and writer.  The existing approach of disassociating and
associating does not work with the new and improved handling of crypto
tokens.  Specifically, the association does not result in the new
shared secret being installed.

Solution
--------

After disassociating, remove the crypto tokens for the remote volatile
reader and writer so that the call to associate will install the new
shared secret.