Durable replay not complete for nack (round 2)
Problem
-------

Before, the datalink thread was responsible for invoking the callback
that replayed durable data.  Thus, durable replay was done when the
call returned.  Now, the discovery and/or user thread is responsible
for replaying durable data and it might not be finished by the time
the first nack arrives.  This results in a request that is gapped
leading to missed durable data.

Solution
--------

Do not process requests for missing data messages when the reader is
still expecting durable data.  See
f19f4d0f3f533918735d9832b4c241b8455cc932.