Authentication gets into a message shootout
Problem
-------

Suppose a request authenticates and sends the final.  The replier gets
the final to complete authentication.  If a message is duplicated or
resent for any reason, then the receiving participant will resend,
causing the other to resend, etc.

Solution
--------

Break the symmetry by only allow the request to resend when
authentication is complete.