Participant location updates ignored
Problem
-------

Participant location updates are enqueued and processed but not
published until the participant appears in the BITs.  This is
problematic when using discovery since the participant will only
appear in the BITs after secure discovery.  Consquently, the location
may never get published or get published with an incorrect mask.

Solution
--------

Don't process the queue until the participant appears in the BITs.