Proper Fix for Missing Thread Status Samples
When using internal thread status BIT with conditions it's possible for
one publish thread status to knock out the previous one before it can be
read because the history depth is one and they are in the same instance.
This tries to fix that by making all the fields a key so they are
different instances, so samples shouldn't be lost in this way.

Removed previous attempt at a fix.