Converting IDL data types to/from JSON requires special codegen
Problem
-------

The existing approach to converting an IDL datatype to/from JSON is
based on code generated by the IDL compiler that is specific to the
JSON conversion.

Solution
--------

Define the ValueReader interface and a JsonValueReader implementation.
(ValueWriter and JsonValueWriter already existed.)  The
ValueReader/ValueWriter interfaces present a generic interface for
deserializers/serializers.  The IDL compiler has the responsiblity of
define generic procedures that use these interfaces to populate a
value (vread) or recite a value (vwrite).

This change creates a number of short-term and long-term
opportunities:

1. The rapidjson pass can be removed from the IDL compiler.

2. The v8 pass can be removed by implementing ValueReader and
ValueWriter for the v8 environment.  Presumably, these implementations
would reside in node-opendds repository.

3. Support for CDR/XCDR (de)serialization can be converted to the
ValueReader/ValueWriter interfaces.

4. (3) supposes manual implemenations of vread and vwrite for type
objects and type identifiers.  These would correspond to the existing
serializer support.

4. A C++ ValueWriter implementation can be defined and used to
generate C++ code that re-creates a value.  This could be used to
great effect in the type object pass to generate the type object and
type identifier output.

5. JsonValueWriter could be applied to type objects and type
identifiers to create an IDL to JSON pass to replace ITL.  Json schema
support may be interesting to consider as well.

Co-authored-by: Fred Hornsey <fred@hornsey.us>