Volatile reader may go from leader to leader
Problem
-------

For background see #2338.  A reader that acks beyond its expected max
sequence number may already be a leader.  The existing code only
removes it from the lagger set so it will be inserted twice into the
leader SNRIS and cause an assert.

Solution
--------

Generalize code so that the reader is removed from the correct SNRIS.