RTPS lacks flow control
Problem
-------

The current implementation of reliability for RTPS is timer driven.
When the timer fires, heartbeats are sent to the readers.  The readers
respond with acknacks with nacks leading to data being resent.  If the
network delay exceeds the heartbeat timer period, then multiple
heartbeats and nacks can be in flight simultaneously.  This causes
data to be resent multiple times which leads to further delays in
processing.

Solution
--------

Record the readers that caused non-final heartbeats to be sent and
measure the delay of the network as the time required to receive an
acknack from all of the readers.  If all the acknacks were received,
recompute the heartbeat period as the measured delay increased by a
safety factor.  Otherwise, increase the heartbeat period by a backoff
factor.