acknack counter is superfluous
Problem
-------

A planned enhancement to flow control is for RtpsWriters to correlate
acknacks with heartbeats.  It will do this using the assumption that
the (OpenDDS) reader will use the heartbeat count for the acknack
count.

Solution
--------

Remove the acknack counter.  This required extending the immediate
heartbeat reply for normal operation to static discovery.