RTPS messages are not logged
Problem
-------

In some situations, it is impossible to collect a packet capture for
debugging RTPS issues.  In this case, it is desirable to have a log of
the RTPS messages.

Solution
--------

Add support for logging RTPS messages.  Payloads for DATA and
DATA_FRAG submessages are not logged.  Messages are logged before any
encryption is applied and after all decryptions are applied.