Disabling relay does not clean up BITs
Problem
-------

The `use_rtps_relay` method in Spdp interpretted the flag
incorrectly, i.e., disabled when the flag was true but attempted to
clean the ParticipantLocation topic.

Calling `use_rtps_relay(false)` in both Spdp and the RTPS transport
does nothing to cleanup the ConnectionRecords that the relay stun task generates.

Solution
--------

Fix the logical errors in `use_rtps_relay` and clean up the
ConnectionRecords when disabling the relay.