Nack filterting is too strict
Problem
-------

When readers reflect the heartbeat count in acknacks, then the counts
are used to filter acknacks that came from historic heartbeats.  This
throws away acknacks that potentially ack data and could prevent a
resend.

Solution
--------

Always apply a valid acknack but only schedule a response if the
acknack is new enough.