Delayed duplicate acknack leads to invalid heartbeat
Problem
-------

Assume that two peers are local but also using the RtpsRelay.  A
preassociation acknack can be delivered after a postassociation
acknack.  This may be interpretted as a reset by the writer.  For
normal RTPS writers, this does not present a problem as it sends
another heartbeat and eventually gets the correct ack.

For the volatile writer in security, it presents a problem since the
value used for the heartbeat last is tracked separately.  If the data
has already been acknowledged, then the writer will send an invalid
heartbeat.  The reader will ignore the invalid heartbeat.  The writer
may keep sending the invalid heartbeat.

Solution
--------

Adjust the heartbeat for the volatile writer so that it is correct.