Zero-count acknacks may cause reset
Problem
-------

For background, see #2537.  A problem with the initial solution is
that resets may still occur due to the way zero-count acknacks are
handled.

A zero-count acknack is used for preassociation when reflecting
heartbeat counts.  Delivering a duplicate zero-count acknack late
causes a reset.

Solution
--------

Provide special handling for zero-count acknacks.  Specifically, they
need to be processed but they should only be used for completing
association.

This commit also resets the durable data to avoid a memory leak.