SEDP modifies transport inst after it is configured
Problem
-------

The call to `reactor_task` in `Sedp::init` causes the transport to be
configured.  `opendds_discovery_default_listener_` and
`opendds_discovery_guid_` are set later so they may or may not be
picked up when the transport is configured.

This causes a delay when using the relay since the local prefix is not
known until later.

Solution
--------

Set `opendds_discovery_default_listener_` and
`opendds_discovery_guid_` before the transport is configured.