ICE password resets accumulate memory when disconnected from relay
Problem
-------

The ICE endpoint manager periodically changes the password.  This
causes new samples to be written to SEDP. If a participant is using
the relay but disconnected, these samples will be buffered.  (The long
term fix is for SEDP to replace the existig sample.)  Additionally,
these password changes generate useless traffic for the relay when
connected.

Solution
--------

Don't change the ICE password unless there are active checklists.