RtpsRelay doesn't publish statistics immediately
Problem
-------

The RtpsRelay defers writing statistics until a new sample needs to be
written and sufficient time has elapsed for writing that sample.  This
means that relay instances will not write statistics samples until
there is traffic.

A load balancer for the RtpsRelay may wish to use these statistics.
If they are unavailable, the load balancer may incorrectly assume that
no instances exist.

This leads to a startup deadlock.

Solution
--------

Write statistics on startup.