RtpsRelay is not scalable
Problem
-------

The existing approach to forwarding messages in the relay relies on
having a complete list of all the readers and writers using the relay,
their associations, and a record indicating which relay is servicing
the associated participant.  This is a large amount of durable data.
Adding a new relay instance means that it must receive all of this
data from its peers.  This places extra load on the existing relay
instances to the point that they start thrashing.

Solution
--------

Reduce the amount of data that must be shared among the relays.  The
new topics are:

* Relay Partitions - Each relay publishes the set of partitions
  corresponding to its clients.

* Relay Partitions Increment - Each relay publishes an additional set
  of partitions for new clients.  The partitions plus the increment
  forms the set of partitions that the relay is servicing.  The
  motivation behind the increment is that it is a smaller and faster
  message which allows all of the relays to be more responsive to new
  clients.  This approach errs on the side of forwarding too many
  messages instead of not sending enough.

* Relay Addresses - Each relay publishes its horizontal addresses.

* SPDP Replay - Each relay publishes requests for replaying SPDP
  messages for new client as they appear.  Currently, this topic is
  reliable but it may make sense to change it to best effort.

Notes
-----

The capability for matching with partition wildcards has been
removed.  It should be restored before merging to master.

The capability for matching based on topic has been removed.  It could
be restored in one of two ways.  First, instead of publishing
partitions, the relay would publish (partition, topic) pairs for
roughly P x T entries.  Second, the topics could be published
independently for roughly P + T entries and a message is forwarded to
a relay if it service at least one source partition and at least one
source topic.  However, most relays will most likely have all of the
topics so the topic would not actually filter any sends.