Listeners not invoked when remote participant times out
Problem
-------

Suppose a participant dies abruptly.  Eventually, all of the readers
associated with the writers of that participant will become
disassociated.  Readers with listeners should receive
`on_data_available` callbacks where the instance state is not
`NOT_ALIVE_NO_WRITERS_INSTANCE_STATE`.  Currently, this does not
happen.

Solution
--------

When a writer is removed, write a synthetic unregister sample for each
instance in the writer.