Network failures destabilize RtpsRelay
Problem
-------

A transient network failure can lead to situations where discovery
purges a remote participant while the participant is still viewed as a
valid network client.  The relay will attempt to rediscover the
participants.  However, there may be too many to rediscover in
parallel which may cause network saturation, excessive CPU
utilization, etc.

Solution
--------

Remove a participant from the list of clients when it is purged from
discovery.  This forces the client to appear to be new and go through
admission control.