Admission control not gated by discovery
Problem
-------

Currently, the dynamic aspect of admission control is governed by time
where the relay will only accept so many new participants per second.
The motivation for doing so is that the relay can discovery so many
participants per second.  This can be done directly by allowing the
discovery of a participant to gate the admission control.

Solution
--------

Keeping a set of pending guids which is increased when a new
participant is admitted and decreased when a participant is
discovered.  One problem with this approach is that there is no way to
say that discovery is completely done, i.e., publications and
subscriptions may come in later.