RtpsRelay RelayHandler pending set fills up
Problem
-------

The pending set is only decreased if a participants timesout in
discovery or activity is recorded which assumes a successful message
parse.  If the pending set fills up, new mesages will be ignored and
the system will effectively be deadlocked.

Solution
--------

Try to free up space in the pending queue early in message processing.