Max directed and undirected gain are difficult to compute
Problem
-------

The motivation behind the maximum directed and undirected gain was to
monitor the number of times a message leads to another message being
sent.  This was measured at the participant, handler, and relay level.

With the switch to partition based routing, computing the actual gains
is difficult and would most likely result in a rough approximation.
Furthermore, participant statistics are most likely not practical to
compute at scale.

Solution
--------

Gain will be measured at the handler and relay level.  For each
received message, the handler will compute the number of sent
messages.

The main piece of information that is lost is the ability to pinpoint
the participant that is causing the gain.  However, one would probably
be able to infer this based on the design of the system, that is, find
the participant that belongs to the most partitions and/or find the
largest partition.