Horizontal forwarding in RtpsRelay has size constraint
Problem
-------

Horizontal sends in the RtpsRelay add a header containing the
partitions and destination guids of the message.  This added header
may make the message too large to send.

Solution
--------

Report an error if the message is larger than the maximum message
size.

Note
----

A permanent fix to this problem is to use something like TCP between
relays that would allow transferring messages of artbitrary size.
Users can also fix the problem by reducing the maximum message size to
leave space for the header added by the relay.