"Proxy" code duplicated between SingleSendBuffer and GuidAddrSet
Problem
-------

SingleSendBuffer and GuidAddrSet have the same problem in that some
sequences of operations on an instance must be atomic.  An operation
involving iterators is one example where the methods used to acquire
the iterators and the code that uses the iterators should be atomic.

Solution
--------

The starting point for a solution is to expose the mutex and create a
scoped guard (ACE_GUARD).  A reference to the object and the guard can
then be passed to functions that are part of the atomic transaction.
However, this leaves room for error as methods can be invoked without
a guard in place.

The first improvement is to combine the reference and guard into a
single object.  This combination is called a Claim and proves that the
object is locked.

The second improvement is to factor out the mutex.  Let T be the class
that needs to be protected.  The template Claimable<T> declares an
instance of T called the core and the mutex used to guard the core.
The core type T must be copyable.

The third improvement is to force all access to the core to go through
a Claim.  A Claimable is "locked" by declaring a Claim.  The Claim
implements indirection `->` so that code with access to the Claim can
invoke methods on the core of the Claimable.  All access to the core
is through the Claim.

Note
----

SingleSendBuffer is not copyable so the Claimable pattern was not
applied to it.