Fix `configure --wireshark-cmake` on Linux
On Linux when using `--wireshark-cmake`, `--wireshark-lib` defaults to
empty. Between making that option and now, a check was added for
`%optdep` locations that require them not to be empty. This added a
field to `%optdep` that allows the value to be empty.

As part of this I also rewrote `%optdep` to use hashes instead of
arrays.