TestFramework lacks generic support for ini files
Problem
-------

The test framework recognizes arguments like "rtps_disc" and
different transport names.  From these arguments, it starts the
InfoRepo if necessary and picks a ini file.

What is desired is a way to specify the exact ini file and let the
test framework determine the discovery mechanism and transport.

Solution
--------

Add support for arguments of the form ini=discovery_transport.ini.
The ini file mentioned will be used when starting the test executables
and the discovery and transport will be inferred from the file name.

Note
----

If necessary, we can extend the convention to
ini=discovery_transport_ignored.ini to support multiple
configurations.